refactor: add Page, Header, NavBar components
rename Header -> Navbar
