+++ src/components/Dashboard.tsx
@@ -1,6 +1,8 @@
 import Editor from '@monaco-editor/react';
-import { useRef } from 'react';
+import React, { useRef } from 'react';
 import useProcessInterval from '../api/submissions';
+import Header from './Header';
+import Page from './Page';
 import Result from './Result';
 
 export default function Example() {
@@ -29,15 +31,10 @@ export default function Example() {
   }
 
   return (
-    <>
-      <header className="bg-white shadow">
-        <div className="max-w-7xl mx-auto py-6 px-4 sm:px-6 lg:px-8">
-          <h1 className="text-3xl font-bold text-gray-900">Dashboard</h1>
-        </div>
-      </header>
+    <Page>
+      <Header title="Dashboard" />
       <main>
         <div className="max-w-7xl mx-auto py-6 sm:px-6 lg:px-8">
-          {/* Replace with your content */}
           <div className="px-4 py-6 sm:px-0">
             <div className="border-4 border-dashed border-gray-200 rounded-lg h-96">
               <Editor
@@ -48,7 +45,6 @@ export default function Example() {
               />
             </div>
           </div>
-          {/* /End replace */}
           <button
             className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
             onClick={() => {
@@ -61,6 +57,6 @@ export default function Example() {
           {result && result}
         </div>
       </main>
-    </>
+    </Page>
   );
 }
